// types/supabase.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "13.0.4"
  }
  public: {
    Tables: {
// ...
      humor_flavor_steps: {
        Row: {
          created_datetime_utc: string
          description: string | null
          humor_flavor_id: number
          humor_flavor_step_type_id: number
          id: number
          llm_input_type_id: number
          llm_model_id: number
          llm_output_type_id: number
          llm_system_prompt: string | null
          llm_temperature: number | null
          llm_user_prompt: string | null
          order_by: number
        }
        Insert: {
          created_datetime_utc?: string
          description?: string | null
          humor_flavor_id: number
          humor_flavor_step_type_id: number
          id?: number
          llm_input_type_id: number
          llm_model_id: number
          llm_output_type_id: number
          llm_system_prompt?: string | null
          llm_temperature?: number | null
          llm_user_prompt?: string | null
          order_by: number
        }
        Update: {
          created_datetime_utc?: string
          description?: string | null
          humor_flavor_id?: number
          humor_flavor_step_type_id?: number
          id?: number
          llm_input_type_id?: number
          llm_model_id?: number
          llm_output_type_id?: number
          llm_system_prompt?: string | null
          llm_temperature?: number | null
          llm_user_prompt?: string | null
          order_by?: number
        }
        Relationships: [
          {
            foreignKeyName: "humor_flavor_steps_humor_flavor_id_fkey"
            columns: ["humor_flavor_id"]
            isOneToOne: false
            referencedRelation: "humor_flavors"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "humor_flavor_steps_humor_flavor_step_id_fkey"
            columns: ["humor_flavor_step_type_id"]
            isOneToOne: false
            referencedRelation: "humor_flavor_step_types"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "humor_flavor_steps_llm_input_type_id_fkey"
            columns: ["llm_input_type_id"]
            isOneToOne: false
            referencedRelation: "llm_input_types"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "humor_flavor_steps_llm_model_id_fkey"
            columns: ["llm_model_id"]
            isOneToOne: false
            referencedRelation: "llm_models"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "humor_flavor_steps_llm_output_type_id_fkey"
            columns: ["llm_output_type_id"]
            isOneToOne: false
            referencedRelation: "llm_output_types"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
      humor_flavors: {
        Row: {
          created_datetime_utc: string
          description: string | null
          id: number
          slug: string
        }
        Insert: {
          created_datetime_utc?: string
          description?: string | null
          id?: number
          slug: string
        }
        Update: {
          created_datetime_utc?: string
          description?: string | null
          id?: number
          slug?: string
        }
        Relationships: []
      }
// ...
      profiles: {
        Row: {
          created_datetime_utc: string | null
          email: string | null
          first_name: string | null
          id: string
          is_in_study: boolean
          is_matrix_admin: boolean
          is_superadmin: boolean
          last_name: string | null
          modified_datetime_utc: string | null
        }
        Insert: {
          created_datetime_utc?: string | null
          email?: string | null
          first_name?: string | null
          id: string
          is_in_study?: boolean
          is_matrix_admin?: boolean
          is_superadmin?: boolean
          last_name?: string | null
          modified_datetime_utc?: string | null
        }
        Update: {
          created_datetime_utc?: string | null
          email?: string | null
          first_name?: string | null
          id?: string
          is_in_study?: boolean
          is_matrix_admin?: boolean
          is_superadmin?: boolean
          last_name?: string | null
          modified_datetime_utc?: string | null
        }
        Relationships: []
      }
// ...
    }
// ...
  }
}